v0.9.7 ~The code has been restructured and minor errors have been fixed. ~Update aiohttp to 3.8.5, emoji to 2.7.0, disnake to 2.9.0.

+++ cogs/commands/distribution.py
@@ -8,7 +8,7 @@ from utils.logger import logger
 
 
 class DistributionCommands(commands.Cog):
-    def __init__(self, bot: commands.Bot):
+    def __init__(self, bot: commands.InteractionBot):
         self.bot = bot
 
     @commands.slash_command(
@@ -19,17 +19,13 @@ class DistributionCommands(commands.Cog):
     async def set_default_role(self, inter: disnake.ApplicationCommandInteraction, role: disnake.Role):
         """
         Обновляет связку роль - реакция на сервере.
-
-        :param inter:
-        :param role:
-        :return:
         """
 
         if role.is_default():
             db.update_guild_settings(inter.guild_id, role_id='NULL')
             await inter.response.send_message('Роль по умолчанию сброшена.', ephemeral=True)
 
-            logger.info(f'[CALL] <@{inter.author.id}> /set_default_role role`s dell')
+            logger.info(f'[CALL] <@{inter.author.id}> /set_default_role role`s del')
 
         elif not role.is_assignable():
             await inter.response.send_message('Невозможно настроить роль по умолчанию, данная роль превосходит роль '
@@ -51,15 +47,15 @@ class DistributionCommands(commands.Cog):
     async def distribution_new_message(self, inter: disnake.ApplicationCommandInteraction):
         """
         Отправляет новое сообщение с автовыдачей ролей по эмодзи.
-
-        :param inter:
-        :return:
         """
 
-        await inter.response.defer(ephemeral=True)
-        await inter.delete_original_response()
+        if not inter.channel.permissions_for(inter.guild.me).add_reactions:
+            await inter.response.send_message('Я не могу ставить реакции в данном канале.', ephemeral=True)
 
-        emb = disnake.Embed(title='Выберите роли',
+            logger.info(f'[CALL] <@{inter.author.id}> /distribution_new_message forbidden add_reactions')
+            return
+
+        emb = disnake.Embed(title='Выбери роли',
                             description='Поставь реакцию для получения соответствующей роли.',
                             colour=disnake.Colour.gold())
 
@@ -67,12 +63,14 @@ class DistributionCommands(commands.Cog):
         for role, reaction in roles:
             emb.add_field(name='', value=f'{reaction} - <@&{role}>', inline=False)
 
-        mes = await inter.channel.send(embed=emb)
+        await inter.response.send_message(embed=emb)
+        mes = await inter.original_message()
 
         for _, reaction in roles:
             await mes.add_reaction(reaction)
 
         db.update_guild_settings(inter.guild_id, distribution=(inter.channel_id, mes.id))
+
         logger.info(f'[CALL] <@{inter.author.id}> /distribution_new_message')
 
     @commands.slash_command(
@@ -84,14 +82,14 @@ class DistributionCommands(commands.Cog):
                                     reaction: str):
         """
         Обновляет связку роль - реакция на сервере.
-
-        :param inter:
-        :param role:
-        :param reaction:
-        :return:
         """
 
-        if role.is_default() or not role.is_assignable():
+        if not inter.channel.permissions_for(inter.guild.me).add_reactions:
+            await inter.response.send_message('Я не могу ставить реакции в данном канале.')
+
+            logger.info(f'[CALL] <@{inter.author.id}> /distribution_add_role forbidden add_reactions')
+
+        elif role.is_default() or not role.is_assignable():
             await inter.response.send_message('Невозможно добавить связку роль - реакция, данная роль превосходит роль '
                                               'бота или является ролью по умолчанию.', ephemeral=True)
 
@@ -101,7 +99,9 @@ class DistributionCommands(commands.Cog):
             if db.update_reaction4role(inter.guild_id, role.id, reaction) == -1:
                 await inter.response.send_message(
                     'Невозможно добавить связку роль - реакция, данная реакция или роль уже участвует в раздаче, '
-                    'вы можете удалить лишние связки роль-реакция командой /distribution_dell_role', ephemeral=True)
+                    'вы можете удалить лишние связки роль-реакция командой /distribution_del_role', ephemeral=True)
+
+                logger.info(f'[CALL] <@{inter.author.id}> /distribution_add_role role or reaction is used')
                 return
 
             logger.info(f'[CALL] <@{inter.author.id}> /distribution_add_role role`s append')
@@ -109,11 +109,11 @@ class DistributionCommands(commands.Cog):
             info = db.get_data('guilds', id=inter.guild_id)[0][4:6]
             if not all(info):
                 await inter.response.send_message(
-                    'Связка роль - реакция добавлена, используйте /distribution_new_message, '
+                    'Связка роль - реакция добавлена, используй /distribution_new_message, '
                     'чтобы создать сообщение выдачи ролей.', ephemeral=True)
                 return
 
-            emb = disnake.Embed(title='Выберите роли',
+            emb = disnake.Embed(title='Выбери роли',
                                 description='Поставь реакцию для получения соответствующей роли',
                                 colour=disnake.Colour.gold())
 
@@ -121,14 +121,18 @@ class DistributionCommands(commands.Cog):
             for role, react in roles:
                 emb.add_field(name='', value=f'{react} - <@&{role}>', inline=False)
 
-            mes = self.bot.get_channel(info[0]).get_partial_message(info[1])
+            try:
+                mes = await self.bot.get_channel(info[0]).get_partial_message(info[1]).edit(embed=emb)
+            except (disnake.errors.NotFound, disnake.errors.Forbidden):
+                await inter.response.send_message(
+                    'Связка роль - реакция добавлена, старое сообщение с раздачей недоступно, '
+                    'используй /distribution_new_message, чтобы создать сообщение выдачи ролей.', ephemeral=True)
+            else:
+                await mes.add_reaction(reaction)
 
-            await mes.edit(embed=emb)
-            await mes.add_reaction(reaction)
-
-            await inter.response.send_message(
-                'Связка роль - реакция добавлена, сообщение с выдачей реакций обновлено.', ephemeral=True
-            )
+                await inter.response.send_message(
+                    'Связка роль - реакция добавлена, сообщение с выдачей реакций обновлено.', ephemeral=True
+                )
 
             logger.info(f'[CALL] <@{inter.author.id}> /distribution_add_role distribution`s update')
 
@@ -146,11 +150,13 @@ class DistributionCommands(commands.Cog):
     async def distribution_del_role(self, inter: disnake.ApplicationCommandInteraction, role: disnake.Role):
         """
         Удаляет связку роль - реакция на сервере.
-
-        :param inter:
-        :param role:
-        :return:
         """
+
+        if not inter.channel.permissions_for(inter.guild.me).add_reactions:
+            await inter.response.send_message('Я не могу ставить реакции в данном канале.')
+
+            logger.info(f'[CALL] <@{inter.author.id}> /distribution_del_role forbidden add_reactions')
+            return
 
         reaction = db.get_data('reaction4role', 'reaction', guild_id=inter.guild_id, role=role.id)
         db.delete_date('reaction4role', guild_id=inter.guild_id, role=role.id)
@@ -159,40 +165,47 @@ class DistributionCommands(commands.Cog):
             reaction = reaction[0][0]
         else:
             await inter.response.send_message(
-                'Данной роли нет в раздаче, используйте /distribution_add_role, '
-                'чтобы добавить роль к раздаче.', ephemeral=True)
+                'Данной роли нет в раздаче, используй /distribution_add_role, '
+                'чтобы добавить роль к раздаче.', ephemeral=True
+            )
+
+            logger.info(f'[CALL] <@{inter.author.id}> /distribution_del_role this role isn`t used')
             return
 
-        logger.info(f'[CALL] <@{inter.author.id}> /distribution_del_role')
+        logger.info(f'[CALL] <@{inter.author.id}> /distribution_del_role role is deleted')
 
         info = db.get_data('guilds', id=inter.guild_id)[0][4:6]
         if not all(info):
             await inter.response.send_message(
-                'Связка роль - реакция удалена, используйте /distribution_new_message, '
+                'Связка роль - реакция удалена, используй /distribution_new_message, '
                 'чтобы создать сообщение выдачи ролей.', ephemeral=True)
             return
 
-        emb = disnake.Embed(title='Выберите роли',
-                            description='Поставьте реакцию для получения соответствующей роли',
+        emb = disnake.Embed(title='Выбери роли',
+                            description='Поставь реакцию для получения соответствующей роли',
                             colour=disnake.Colour.gold())
 
         roles = db.get_data('reaction4role', 'role, reaction', guild_id=inter.guild_id)[:10]
         for role, react in roles:
             emb.add_field(name='', value=f'{react} - <@&{role}>', inline=False)
 
-        mes = self.bot.get_channel(info[0]).get_partial_message(info[1])
+        try:
+            mes = await self.bot.get_channel(info[0]).get_partial_message(info[1]).edit(embed=emb)
+        except (disnake.errors.NotFound, disnake.errors.Forbidden):
+            await inter.response.send_message(
+                'Связка роль - реакция удалена, старое сообщение с раздачей недоступно, '
+                'используй /distribution_new_message, чтобы создать сообщение выдачи ролей.', ephemeral=True)
+        else:
+            await mes.remove_reaction(reaction, self.bot.user)
 
-        await mes.edit(embed=emb)
-        await mes.remove_reaction(reaction, self.bot.user)
-
-        await inter.response.send_message(
-            'Связка роль - реакция удалена, сообщение с выдачей реакций обновлено.', ephemeral=True
-        )
+            await inter.response.send_message(
+                'Связка роль - реакция удалена, сообщение с выдачей реакций обновлено.', ephemeral=True
+            )
 
         logger.info(f'[CALL] <@{inter.author.id}> /distribution_add_role distribution`s update')
 
 
-def setup(bot: commands.Bot):
+def setup(bot: commands.InteractionBot):
     """Регистрация команд бота."""
 
     bot.add_cog(DistributionCommands(bot))
